% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDFql.R
\docType{package}
\name{RHDFql-package}
\alias{RHDFql}
\alias{RHDFql-package}
\title{RHDFql: Interface to 'HDFql'}
\description{
A DBI-like interface to HDF files using HDFql.
}
\details{
TBD.
}
\author{
\strong{Maintainer}: Michael Koohafkan \email{michael.koohafkan@gmail.com}

}
